requirements.txt
# Video Processing
moviepy==2.2.1
pillow==11.0.0

# AI Integration
anthropic>=0.18.0
google-generativeai>=0.8.0
google-genai>=1.0.0

# YouTube API
google-api-python-client==2.188.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0

# Text-to-Speech
edge-tts==7.2.7

# Storage & Database
supabase>=2.3.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic==2.12.5
pydantic-settings==2.7.0
pydantic-extra-types==2.10.1
pyyaml==6.0.2

# Email
resend>=2.0.0

# Retry logic
backoff>=2.2.0

# Trend Discovery
pytrends>=4.9.0

# Utilities
requests>=2.31.0
python-dateutil==2.9.0
httpx==0.28.1
tuspy==1.1.0
